+++ app/api/api_v1/routers/genes.py
@@ -1,10 +1,12 @@
 from fastapi import APIRouter, Query, Request, Depends, Response, encoders
 from typing import List, Optional
 import pandas as pd
+import io
+from starlette.responses import StreamingResponse
 
 from app.db.session import get_db
 from app.db.crud import (
-    get_genes, get_genes_download, get_genes_by_defense, prepare_csv_file, get_genes_by_cluster, prepare_fasta_file
+    get_genes, get_genes_download, get_genes_by_defense, prepare_csv_file, get_genes_by_cluster
 )
 from app.db.schemas import GeneBase
 
@@ -83,4 +85,28 @@ async def genes_by_cluster(
     else: # selected a fasta file
         return prepare_fasta_file(genes_by_cluster, prot)
 
+
+# prepares a fasta file. returns as a text file to the user. the front end needs to translate it into a .faa file.
+def prepare_fasta_file(df, prot):
+    final_txt = ""
+    for index, row in df.iterrows():
+        locus_tag, start_g, end_g, name, g_accession, cluster_index,product_accession  = row['locus_tag'], row['start'], row['end'], row['name'], row['genomic_accession'], row['cluster_index'], row['product_accession']
+        seq = row['protein_sequence'] if prot else row['dna_sequence']
+        every = 120
+        seq = '\n'.join(seq[i:i+every] for i in range(0, len(seq), every))
+        type = 'prot' if prot else 'dna'
+        newentry = ">{}_{} [locus_tag = {}] [location = {}..{}] [name = {}] [cluster_index = {}] [product_accession = {}] \n{}\n".format(g_accession,type,locus_tag,start_g,end_g,name,cluster_index,product_accession,seq)
+        final_txt += newentry
+
+    output = io.StringIO()
+    output.write(final_txt)
+
+    # Returns a csv prepared to be downloaded in the FrontEnd
+    response = StreamingResponse(iter([output.getvalue()]),
+                                 media_type="text/plain"
+                                 )
+
+    response.headers["Content-Disposition"] = "attachment; filename=export.txt"
+
+    return response
